Update: try to use ducks type
https://github.com/erikras/ducks-modular-redux
