--- app/javascript/src/ProjectPage/actions.js
@@ -1,8 +1,0 @@
-import { PROJECT_PAGE } from 'actionTypes'
-
-export function fetchTasks(page = 1) {
-  return {
-    type: PROJECT_PAGE.FETCH_TASKS.REQUEST,
-    page,
-  }
-}
