+++ app/javascript/src/ducks/index.js
@@ -1,0 +1,3 @@
+import tasks from './tasks'
+
+export { tasks }
