+++ app/javascript/src/store.js
@@ -2,12 +2,9 @@ import { createStore, combineReducers, applyMiddleware } from 'redux'
 import { composeWithDevTools } from 'redux-devtools-extension'
 import createSagaMiddleware from 'redux-saga'
 import rootSaga from './rootSaga'
+import * as reducers from './ducks'
 
-import projectPage from './ProjectPage/reducer'
-
-const rootReducer = combineReducers({
-  projectPage,
-})
+const rootReducer = combineReducers(reducers)
 const sagaMiddleware = createSagaMiddleware()
 const store = createStore(
   rootReducer,
